Extract data structure to header
