+++ core/internal/ssl_stripping/silent.h
@@ -4,6 +4,8 @@
 #include "PcapLiveDevice.h"
 #include "ssl_stripping/public.h"
 #include "ssl_stripping/ssl_stripping_strategy.h"
+#include <mutex>
+#include <queue>
 #include <utility>
 
 namespace ATK::SSL {
@@ -68,6 +70,18 @@ class SilentSslStrippingStrategy : public ATK::SSL::SslStrippingStrategy {
     void execute() override;
 
   private:
+    // Function to get global variables safely
+    struct HttpMessageData {
+        std::queue<std::string> httpMessages;
+        std::mutex httpMessagesMutex;
+        std::condition_variable httpMessagesCV;
+    };
+
+    static HttpMessageData &getHttpMessageData() {
+        static HttpMessageData data;
+        return data;
+    }
+
     SilentSslStrippingStrategy(pcpp::PcapLiveDevice *device,
                                pcpp::IPv4Address attackerIp,
                                std::vector<pcpp::IPv4Address> victimIps,
